feat: expose controller config (#554)

+++ common/src/main/java/com/automq/rocketmq/common/config/ControllerConfig.java
@@ -43,7 +43,7 @@ public interface ControllerConfig extends GrpcClientConfig {
         return 10;
     }
 
-    default int nodeAliveIntervalInSecs() {
+    default long nodeAliveIntervalInSecs() {
         return 60;
     }
